set info --env command
